backend/requirements.txt
# Backend Python dependencies
fastapi
uvicorn
requests
PyPDF2
python-dotenv
llama-api-client